Shared validation logic tweaking
